Fix the flicker problem on iOS (Safari)
In Safari, the TinySlider has a bug that when swiping images flicker or are displayed only half. The problem can be fixed with the attached change. 
In the core of the TinySlider the whole thing was not fixed until today, therefore the fix here. 

There are several open issues.
Example: https://github.com/ganlanyuan/tiny-slider/issues/683